Fix edited comment not showing notifications
