+++ client/coral-embed-stream/src/tabs/stream/containers/CommentBox.js
@@ -7,31 +7,10 @@ import { can } from 'coral-framework/services/perms';
 import Slot from 'coral-framework/components/Slot';
 import { connect } from 'react-redux';
 import { CommentForm } from '../components/CommentForm';
+import { notifyForNewCommentStatus } from '../helpers';
 
 // TODO: change this...
 export const name = 'talk-plugin-commentbox';
-
-const notifyReasons = ['LINKS', 'TRUST'];
-
-function shouldNotify(actions = []) {
-  return actions.some(
-    ({ __typename, reason }) =>
-      __typename === 'FlagAction' && notifyReasons.includes(reason)
-  );
-}
-
-// Given a newly posted comment's status, show a notification to the user
-// if needed
-export const notifyForNewCommentStatus = (notify, comment, actions) => {
-  if (comment.status === 'REJECTED') {
-    notify('error', t('comment_box.comment_post_banned_word'));
-  } else if (
-    comment.status === 'PREMOD' ||
-    (comment.status === 'SYSTEM_WITHHELD' && shouldNotify(actions))
-  ) {
-    notify('success', t('comment_box.comment_post_notif_premod'));
-  }
-};
 
 /**
  * Container for posting a new Comment
@@ -87,7 +66,7 @@ class CommentBox extends React.Component {
         // Execute postSubmit Hooks
         this.state.hooks.postSubmit.forEach(hook => hook(data));
 
-        notifyForNewCommentStatus(notify, postedComment, actions);
+        notifyForNewCommentStatus(notify, postedComment.status, actions);
 
         if (commentPostedHandler) {
           commentPostedHandler();
